Added all of the necessary JS code to add and remove guests
This included a significant amount of coffeescript "boilerplate"
as well as an overhaul of the controllers to use view models.
These view models can use some additional love in the near future,
but for now they serve as a decent enough abstraction for containing
rendering logic; especially for consistent data representations.