Fixing other/previous guests when adding new guests on index view
When adding a guest to an event on the index page the other guest
lists of the other invitations were not being updated.  This commit
fixes that issue by adding an altogether way too convoluted process
to merge new unique guests into those lists.  It does so in a pretty
awful way right now but it works and I'll refactor it later.