Fixing guest list completed task issue
I don't have a good way for guests to indicate that they don't
have any guests to add, so the progress bar can't really indicate
whether that task is completed.  Doh.  So for now we won't
include it in the tasks and hope people just do it anyway
with prompting via emails.